change PeoplePage to show message if table is empty
